# public/api/openapi.yaml
openapi: 3.1.0
info:
  title: Phototracker API
  version: 0.1.0
  license:
    name: Proprietary
  description: |
    API for tracking film photographs — cameras, lenses, film stocks, development profiles, film holders, rolls, exposures, and reference images.
servers:
  - url: https://darkcloth.zone
    description: Primary production hostname.
  - url: https://phototracker.graha.ms
    description: Legacy/staging hostname during the transition.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    User:
      type: object
      required: [id, email, default_timezone, auto_use_current_location, created_at, updated_at]
      properties:
        id:
          type: string
          description: ULID
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        email:
          type: string
          format: email
          example: "user@example.com"
        default_timezone:
          type:
            - string
            - "null"
          description: Stored IANA timezone preference. Null when unset.
          example: "America/Denver"
        auto_use_current_location:
          type: boolean
          description: Whether new photo logs should request browser geolocation automatically.
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-04-29T20:26:05Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-04-29T20:26:05Z"
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      operationId: register
      summary: Create a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                type: object
                required: [token, user]
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/login:
    post:
      operationId: login
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                type: object
                required: [token, user]
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/me:
    get:
      operationId: getMe
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMe
      summary: Update current authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: New email address. Changing email requires current_password.
                current_password:
                  type: string
                  minLength: 1
                  description: Required when email is present.
                default_timezone:
                  type:
                    - string
                    - "null"
                  description: IANA timezone name or null to clear the preference.
                auto_use_current_location:
                  type: boolean
                  description: Whether new photo logs should request browser geolocation automatically.
              dependentRequired:
                email:
                  - current_password
      responses:
        "200":
          description: Current user updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/password:
    patch:
      operationId: updatePassword
      summary: Change the current password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                  minLength: 1
                new_password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Gear: Cameras ─────────────────────────────────────────────────────────────
...